♻️ refactor: WeatherService 구조 정리 및 record 기반으로 리팩토링
